server/requirements.txt
asgiref==3.9.1
channels==4.3.1
channels_redis==4.3.0
click==8.2.1
colorama==0.4.6
Django==5.2.5
django-cors-headers==4.7.0
django-redis==6.0.0
djangorestframework==3.16.1
h11==0.16.0
msgpack==1.1.1
psycopg2-binary==2.9.10
python-dotenv==1.1.1
redis==6.4.0
sqlparse==0.5.3
tzdata==2025.2
uvicorn==0.35.0
websockets==15.0.1
